Fix various imports due to restructure
